make ledalerts self contained
