+++ EBike_wireless_TSDZ2/firmware/include/ledalert.h
@@ -141,3 +141,4 @@ volatile uint8_t ui8_led_red_intensity;
 volatile uint8_t ui8_led_green_intensity;
 volatile uint8_t ui8_led_blue_intensity;
 volatile uint8_t ui8_led_on;
+
